requirements.txt
PyYAML==5.1.2
cx_oracle==7.2.3
numpy==1.17.2
pandas==0.25.0
geopandas==0.5.1
shapely==1.6.4.post2
rtree==0.8.3
simplekml
pysftp
paramiko
